+++ Examples/NotificationsTest/ViewController.h
@@ -1,6 +1,6 @@
 //
 //  ViewController.h
-//  NotificationsTesr
+//  NotificationsTest
 //
 //  Created by Nick Lockwood on 20/11/2013.
 //  Copyright (c) 2013 Charcoal Design. All rights reserved.