+++ FXNotifications/FXNotifications.h
@@ -1,7 +1,7 @@
 //
 //  FXNotifications.h
 //
-//  Version 1.0
+//  Version 1.0.1
 //
 //  Created by Nick Lockwood on 20/11/2013.
 //  Copyright (c) 2013 Charcoal Design