+++ RELEASE NOTES.md
@@ -1,3 +1,9 @@
+Version 1.0.1
+
+- Fixed bug where removing the observer didn't work
+- The queue parameter now defaults to [NSOperationQueue currentQueue] if nil
+- Added CocoaPods podspec
+
 Version 1.0
 
 - Initial release